Print()'s argument "nt" should be an integer.

+++ src/Timing.hpp
@@ -122,11 +122,11 @@ namespace timing
     Timer & New_Timer(const std::string &name);
     void Wait(const double seconds);
     void Print_N_Times(const std::string x, const size_t N, const bool newline = true);
-    void _Print(const double nt);
+    void _Print(const uint64_t nt);
 #ifndef DISABLE_TIMING
-    inline void Print(const double nt)  { _Print(nt);               }
+    inline void Print(const uint64_t nt) { _Print(nt);               }
 #else // #ifndef DISABLE_TIMING
-    inline void Print(const double nt)  { /* Don't do anything */   }
+    inline void Print(const uint64_t nt) { /* Don't do anything */   }
 #endif // #ifndef DISABLE_TIMING
 
     // **********************************************************
